Relocate intrinsic modules to src/video/
